requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
httpx>=0.26.0
pydantic-settings>=2.0.0
